requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
gevent==22.10.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pysqlite3-binary==0.5.0
Werkzeug==2.3.4
zope.event==4.6
zope.interface==6.0
